+++ imgView/imgView.cpp
@@ -45,7 +45,7 @@ int CImgView::init(const QString& filePath, const CCfgMgr* cfgMgr)
 
         //图片管理
         //如果返回空指针,代表初始化失败
-        auto dspOrder = (CFG::DspOrder::ByName == cfg->imgView.dspOrder)    //TODO: 在imgMgr类的内部实际处理最终顺序
+        auto dspOrder = (static_cast<int>(CFG::DspOrder::ByName) == cfg->imgView.dspOrder)    //TODO: 在imgMgr类的内部实际处理最终顺序
             ? QDir::SortFlags(QDir::Name | QDir::IgnoreCase)
             : QDir::SortFlags(QDir::Time);
         m_imgMgr = CImgMgrFac::create(filePath, dspOrder);
